+++ src/Interfaces/ResponseInterface.php
@@ -10,6 +10,8 @@ namespace BlueMvc\Core\Interfaces;
 use BlueMvc\Core\Interfaces\Collections\HeaderCollectionInterface;
 use BlueMvc\Core\Interfaces\Collections\ResponseCookieCollectionInterface;
 use BlueMvc\Core\Interfaces\Http\StatusCodeInterface;
+use DataTypes\Interfaces\HostInterface;
+use DataTypes\Interfaces\UrlPathInterface;
 
 /**
  * Interface for Response class.
@@ -113,6 +115,21 @@ interface ResponseInterface
      * @param ResponseCookieCollectionInterface $cookies The cookies.
      */
     public function setCookies(ResponseCookieCollectionInterface $cookies);
+
+    /**
+     * Sets a cookie value.
+     *
+     * @since 1.1.0
+     *
+     * @param string                  $name       The name.
+     * @param string                  $value      The value.
+     * @param \DateTimeInterface|null $expiry     The expiry time or null if no expiry time.
+     * @param UrlPathInterface|null   $path       The path or null if no path.
+     * @param HostInterface|null      $domain     The domain or null if no domain.
+     * @param bool                    $isSecure   True if cookie is secure, false otherwise.
+     * @param bool                    $isHttpOnly True if cookie is http only, false otherwise.
+     */
+    public function setCookieValue($name, $value, \DateTimeInterface $expiry = null, UrlPathInterface $path = null, HostInterface $domain = null, $isSecure = false, $isHttpOnly = false);
 
     /**
      * Sets the expiry time.